move readiness check to integration tests
